Cast to double to solve bug computing number of iterations on large problems.

+++ tconv2.cc
@@ -130,7 +130,7 @@ int main(int argc, char* argv[])
   cout << "mx=" << mx << ", my=" << my << endl;
   
   if(N == 0) {
-    N=N0/(nx*ny);
+    N=(int) ((double)N0/((double) (nx*ny)));
     if(N < 10) N=10;
   }
   cout << "N=" << N << endl;
